Move some code to ert.parsing
