+++ src/ert/_c_wrappers/job_queue/workflow.py
@@ -4,7 +4,8 @@ import time
 from tempfile import mkdtemp
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple
 
-from ert._c_wrappers.config import ConfigParser, ConfigValidationError, UnrecognizedEnum
+from ert._c_wrappers.config import ConfigParser, UnrecognizedEnum
+from ert.parsing import ConfigValidationError
 
 if TYPE_CHECKING:
     from ert._c_wrappers.enkf import EnKFMain
